Move binding inspection and report generation out of internal/cli into internal/inspection so the CLI is a thin adapter over reusable read-only inspection logic.
