+++ internal/cli/inspect.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 
+	"github.com/sonda-red/kleym/internal/inspection"
 	"github.com/spf13/cobra"
 )
 
@@ -11,6 +12,14 @@ var (
 	errInspectBindingHasErrorFindings   = errors.New("binding inspection completed with error findings")
 	errInspectBindingHasWarningFindings = errors.New("binding inspection completed with warning findings in strict mode")
 )
+
+var newBindingInspectionRunner = func(opts *Options) (inspection.BindingInspector, error) {
+	return inspection.NewKubernetesBindingInspector(inspection.Config{
+		Context:    opts.Context,
+		Kubeconfig: opts.Kubeconfig,
+		Timeout:    opts.Timeout,
+	})
+}
 
 // newInspectCommand creates the inspect command group under the root CLI.
 func newInspectCommand(opts *Options) *cobra.Command {
@@ -38,7 +47,7 @@ func newInspectCommand(opts *Options) *cobra.Command {
 			defer cancel()
 
 			report, inspectErr := inspector.InspectBinding(ctx, opts.Namespace, args[0])
-			if err := WriteBindingInspectionReport(cmd.OutOrStdout(), opts.Output, report); err != nil {
+			if err := inspection.WriteBindingInspectionReport(cmd.OutOrStdout(), opts.Output, report); err != nil {
 				return withExitCode(exitInternal, err)
 			}
 			code, err := inspectionExitCode(report, inspectErr, opts.Strict)
@@ -53,20 +62,20 @@ func newInspectCommand(opts *Options) *cobra.Command {
 	return inspectCmd
 }
 
-func inspectionExitCode(report BindingInspectionReport, inspectErr error, strict bool) (int, error) {
+func inspectionExitCode(report inspection.BindingInspectionReport, inspectErr error, strict bool) (int, error) {
 	if inspectErr != nil {
-		if errors.Is(inspectErr, errBindingInspectionNotFound) {
+		if errors.Is(inspectErr, inspection.ErrBindingInspectionNotFound) {
 			return exitBindingNotFound, inspectErr
 		}
-		if errors.Is(inspectErr, errBindingInspectionErrorFindings) {
+		if errors.Is(inspectErr, inspection.ErrBindingInspectionErrorFindings) {
 			return exitInspectionIssue, errInspectBindingHasErrorFindings
 		}
 		return exitUsage, inspectErr
 	}
-	if hasErrorSeverityFinding(report.Findings) {
+	if inspection.HasErrorSeverityFinding(report.Findings) {
 		return exitInspectionIssue, errInspectBindingHasErrorFindings
 	}
-	if strict && hasWarningSeverityFinding(report.Findings) {
+	if strict && inspection.HasWarningSeverityFinding(report.Findings) {
 		return exitInspectionIssue, errInspectBindingHasWarningFindings
 	}
 	return exitOK, nil
