+++ internal/inspection/report_test.go
@@ -1,4 +1,4 @@
-package cli
+package inspection
 
 import (
 	"bytes"
